requirements.txt
einops==0.8.0
numpy==1.26.4
accelerate==0.34.2
huggingface-hub==0.24.6
open-clip-torch==2.26.1
opencv-python==4.6.0.66
gdown==4.4.0
pillow==10.4.0
pyarrow==17.0.0
scipy==1.11.1
timm==1.0.9
tokenizers
torch==2.0.1
torchaudio==2.0.2
torchdata==0.6.1
torchmetrics==1.4.1
torchvision==0.15.2
setuptools==60.2.0
tqdm
transformers==4.35.2
urllib3==1.26.20
xformers==0.0.22
streamlit-keyup==0.2.0
lpips
clean-fid
peft
dominate
diffusers==0.25.1
vision_aided_loss
clip_interrogator==0.6.0
wandb
matplotlib==3.5.3